Changing type of Margin, Border, FitToPage, FontSize, & FontName (#170)
* Revert "Revert "Describe annotations, delete URIType, add FontName, Border, and Margin (#159)" (#163)"

This reverts commit 853ab7cccffbc4ba103bc8d2d65577e589cb925f.

* Update PDF.xml

Margin, Border, FitToPage, FontSize, & FontName are new features of the PDF_API. The annotation shows you if this option is available from the server.
=> Changing the types to boolean.

We have no default FontName, FonstSize and so on. Only the pdf API knows this. If the pdf API add this default knowledge, it could be (for an older FES), that this information is not available. Then we would have to set NULL for these default FontName, FontSize, etc. setting. Therefore, I would not add this here.

* Update PDF.md

* Update PDF.json

* Update PDF.Features-examples.xml

* Update PDF.Features-examples.json

* Remove:

* Remove: Remove "a the" from the description.

Co-authored-by: Ralf Handl <ralf.handl@sap.com>