Inparam (#303)
ValueListParameterIn is of type PropertyPath. Therefore it might contain
a collection-based navigation or structural property. In this case we
don't have only one 'eq' comparison in the filter (of the value list)
but a chain of 'eq' comparisons separarted by 'or' operators. This will
be documented now with that change.