Merge cells in table columns (#324)
Create an opportunity to define columns in a table where the cells are
merged if the adjacent values are the same (in html rowSpan would be
used). I decided to go as a starting point with an HTML5 collection of
propertyPath that can be applied to UI.LineItem. Imo this new annotation
makes only sense for UI.LineItem, and I didn't want to change the term
itself.
Other possibilities would be applying the annotation to properties or
adding something to the UI.PresentationVariant.

---------

Co-authored-by: Heiko Theißen <heiko.theissen@sap.com>