Support annotating common expressions as sort order of entity sets (#333)
This proposal addresses the requirement to use a common expression in
the sort order annotated to an entity set. For this purpose, this
proposal extends the structure type of term `Common.SortOrder`.

**Rationale for this extension**: In the concrete use case of this
requirement, EDMX is used to as an _interface definition language_
(a.k.a. _service provider interface (SPI)_) to describe the sort
behavior a service implementation compliant with the API definition must
offer for the annotated entity set. Allowing the use of an expression
for a sort item mirrors the expressiveness of the protocol ABNF and
leads to a self-contained API description that can be confidently
published on the SAP Business Accelerator Hub (that is, no need for an
out-of-band textual description).

**Discussion needed**: The type of term `Common.SortOrder` is reused in
`UI.PresentationVariant`. If there are concerns that this proposal adds
to much freedom for services and makes it hard for UIs to interpret
`Common.SortOrder` annotations because they lack a common expression
parser, an **alternative design** would be to keep the term
`Common.SortOrder` as it is today and create a new term
`Common.SortOrderExpressions` with a type derived from
`Common.SortOrderType` adding the expression to this structure.

Example from the use case of the SAP Build Work Zone product: Sort
instances of the `entities` entity set such that instances with
`identification/entityType = "Site"` are returned before all other
instances

```xml
<Annotations Target="entities">
  <Annotation Term="Common.SortOrder">
    <Collection>
      <Record>
        <PropertyValue Property="Expression">
          <Eq>
            <Path>identification/entityType</Path>
            <String>Site</String>
          </Eq>
        </PropertyValue>
      </Record>
    <Collection>
  </Annotation>
</Annotations>

  …
</Annotations>

```

---------

Co-authored-by: Ralf Handl <ralf.handl@sap.com>
Co-authored-by: Heiko Theißen <heiko.theissen@sap.com>