Add identifier to action groups (#356)
For the consumers there is a need to identify action groups uniquely.
Similar to CollectionFacets there is nothing that would uniquely
identify this group/collection. If it would be the concatenation of all
entries, already moving or removing one entry would change the ID
(although the semantics of the group doesn't change). If there are
others idea to achieve this unique identification of an action group,
pls. let me know - I would be really interested in.

---------

Co-authored-by: Heiko Theißen <heiko.theissen@sap.com>