Introducing Relative Value List collections (#392)
There are use cases in which applications already have a
navigationProperty to the available items that should be shown in a
value help, for example in a sales order item, I have a "toMaterials"
navigation property that shows all the materials that can be chosen for
this item. Without relative value list collections, applications need to
pass all information via in parameters to the collection. With this
change, applications can use the navigation property directly

---------

Co-authored-by: Heiko Theißen <heiko.theissen@sap.com>